pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "Synthetic-Societies"
version = "0.1.0"
description = ""
authors = [
    { name = "Ang Li",email = "dmisno1@gmail.com" },
    { name = "Prisha Priyadarshini", email = "prishapriya31@gmail.com" }
]
requires-python = ">=3.10, <3.14"

dependencies = [
    "autogen (>=0.10.0,<0.11.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "ag2[gemini] (>=0.10.1,<0.11.0)"
]

[project.urls]
homepage = "https://github.com/ale-mateus/Algoverse---Synthetic-Society"

[tool.poetry]
package-mode = false